+++ CHANGELOG.md
@@ -1,1 +1,5 @@
 # ????-??-??
+
+* Rework bundle config
+  * Make the config option naming less ambiguous
+  * Add an option to specify the local validation time leeway